refactor: moving HomeServices
