+++ src/app/home/components/country-item/country-item.component.ts
@@ -16,7 +16,6 @@ export class CountryItemComponent {
 
   ngOnChanges() {
     if (this.country) {
-      // TODO: CREATE A PRIVATE FUNCTION
       const { name, flags, capital, subregion } = this.country;
       this.flag = flags.png;
       this.name = name.common;
