+++ src/app/services/HomeServices/home.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Regions } from '../shared/interfaces';
+import { Regions } from '../../shared/interfaces';
 
 @Injectable({
   providedIn: 'root',
